Added code to persist data upon application being stopped. Added Timber. Global Code Cleanup
Notes:

onCreate:
    Inflate the activity's UI, whether that's using findViewById or databinding.
    Initialize variables.
    Do any other initialization that only happens once during the activity lifetime.

onStart:
    Start any sensors, animations or other procedures that need to start when the activity is visible.

onResume:
      Start any sensors, animations or other procedures that need to start when the activity has focus (the activity the user is currently interacting with).

onPause:
    Stop any sensors, animations or other procedures that should not run when the activity doesn't have focus and is partially obscured.
    Keep execution fast. The next activity is not shown until this completes.

onStop:
   Stop any sensor, animations or other procedures that should not run when the activity is not on screen.
You can use this to persist (permanently save) data, which you’ll be learning more about in Lesson 6
Stop logic that updates the UI. This should not be running when the activity is off-screen; it's a waste of resources.
There are also restrictions as soon as the app goes into the background,

onDestroy:
    Tear down or release any resources that are related to the activity and are not automatically released for you. Forgetting to do this could cause a memory leak!
