+++ Udacity/ColorMyViews/app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 30
 
     defaultConfig {
-        applicationId "com.tc2r1.colormyviews"
+        applicationId "com.nudennie.colormyviews"
         minSdk 21
         targetSdk 30
         versionCode 1
