+++ Udacity/DessertPusher/README.md
@@ -1,13 +1,17 @@
-# Activity Lifecycle - DessertPusher 
+# Activity Lifecycle - DessertPusher
 
-This is the toy app for lesson 4 of the [Android App Development in Kotlin course on Udacity](https://classroom.udacity.com/courses/ud9012/lessons/e487c600-ed68-4576-a35a-12f211cf032e/concepts/6a155d63-8153-4a56-95cb-1dfdf06aa173).
+This is the toy app for lesson 4 of
+the [Android App Development in Kotlin course on Udacity](https://classroom.udacity.com/courses/ud9012/lessons/e487c600-ed68-4576-a35a-12f211cf032e/concepts/6a155d63-8153-4a56-95cb-1dfdf06aa173)
+.
 
 ## DessertPusher
 
 DessertPusher is a game about making desserts. Press the button, make a dessert, earn the big bucks.
 
-This app is for exploring the [Activity Lifecycle callback methods](https://developer.android.com/guide/components/activities/activity-lifecycle) and the [Lifecycle](https://developer.android.com/reference/android/arch/lifecycle/Lifecycle) object.
-
+This app is for exploring
+the [Activity Lifecycle callback methods](https://developer.android.com/guide/components/activities/activity-lifecycle)
+and the [Lifecycle](https://developer.android.com/reference/android/arch/lifecycle/Lifecycle)
+object.
 
 ## Screenshots
