+++ Udacity/DessertPusher/app/src/main/AndroidManifest.xml
@@ -15,17 +15,17 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.android.dessertpusher">
+    package="com.nudennie.dessertpusher">
 
     <application
-        android:name=".PusherApplication"
+        android:name="com.nudennie.dessertpusher.PusherApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_dessert_pusher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_dessert_pusher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.nudennie.dessertpusher.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
