--- Udacity/DessertPusher/app/src/main/java/com/example/android/dessertpusher/PusherApplication.kt
@@ -1,14 +1,0 @@
-package com.example.android.dessertpusher
-
-import android.app.Application
-import timber.log.Timber
-
-class PusherApplication : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-
-        // Initialize Timber Debugging.
-        Timber.plant(Timber.DebugTree())
-    }
-}
