+++ Udacity/DessertPusher/app/src/main/res/drawable/jellybean.xml
@@ -19,34 +19,34 @@
     android:height="700dp"
     android:viewportWidth="500"
     android:viewportHeight="700">
-  <path
-      android:fillColor="#C2CACC"
-      android:pathData="M451.5,482.8c0,-17 -41.2,-30.8 -92.1,-30.8c-9.3,0 -18.3,0.5 -26.8,1.3c-19.6,-3.3 -41.8,-5.2 -65.3,-5.2c-8,0 -15.9,0.2 -23.6,0.6c-21,-5.2 -46.9,-8.2 -75,-8.2c-70.2,0 -127,19 -127,42.5c0,21.2 46.3,38.7 106.8,42c26.4,12.4 69.8,20.5 118.8,20.5c68.1,0 125.2,-15.6 141.2,-36.8C434.3,503.4 451.5,493.8 451.5,482.8z" />
-  <path
-      android:fillColor="#FF9800"
-      android:pathData="M393.2,495.4c-46.7,-9.5 -87.9,-33.6 -124.8,-69.3c-18.2,-17.5 -22.6,-45 -10.7,-67.3c15.4,-28.7 51.9,-38.6 79.4,-21.3c23,14.4 48.1,25.2 74.4,32.1c42.3,11.2 63.1,58.7 42.5,97.2C442.4,488.7 417.6,500.4 393.2,495.4" />
-  <path
-      android:fillColor="#FFBC41"
-      android:pathData="M405.7,378.1l-0.2,0.1l0.1,0.2c5.2,1.6 10.2,3.8 14.8,6.5c4.7,2.7 9,5.9 12.8,9.7c3.7,3.8 7.2,7.9 10.2,12.3c3.1,4.4 6,9 8.5,13.8l0.2,0.1l0.1,-0.2c-1.2,-5.4 -3.3,-10.6 -6.3,-15.3c-3,-4.7 -6.7,-8.9 -11,-12.4c-4.2,-3.6 -8.8,-6.7 -13.8,-9.2C416.3,381.2 411.1,379.3 405.7,378.1L405.7,378.1" />
-  <path
-      android:fillColor="#FFBC41"
-      android:pathData="M306.2,338c-3.3,0 -6.6,0.3 -9.9,0.9c-4.3,0.8 -8.5,2.1 -12.5,3.9l0,0l-0.1,0.2l0.2,0.1c4.1,-1.4 8.3,-2.3 12.6,-2.7c1.9,-0.2 3.8,-0.3 5.7,-0.3c2.3,0 4.7,0.1 7,0.4c4.2,0.7 8.3,1.6 12.3,2.9c4.1,1.2 8,2.7 12.1,4.3h0.2l0,0l0,-0.2c-3.3,-3 -7.2,-5.3 -11.3,-6.9c-4.2,-1.5 -8.6,-2.3 -13.1,-2.5C308.4,338 307.3,338 306.2,338" />
-  <path
-      android:fillColor="#42A5F5"
-      android:pathData="M96.9,504.7c57.2,8 114,-3 170.6,-28.8c27.9,-12.7 44,-42.4 39.4,-72.7c-6,-39 -43.7,-65 -82.2,-56.4c-32.1,7.2 -65.2,9.4 -98,6.6c-52.8,-4.3 -95.9,41.6 -87.8,93.9C43.4,477.2 67,500.5 96.9,504.7" />
-  <path
-      android:fillColor="#7FC2F8"
-      android:pathData="M120.6,365c-10.2,0 -20.4,1.9 -29.9,5.8c-6.3,2.4 -12.2,5.6 -17.6,9.7c-5.3,4.1 -9.9,9.1 -13.5,14.7v0.2h0.2c4.7,-4.5 9.8,-8.7 15.1,-12.5c5.2,-3.8 10.8,-7.1 16.7,-9.9c6,-2.7 12.2,-4.6 18.6,-5.8c5.2,-1 10.4,-1.4 15.6,-1.4c1.3,0 2.7,0 4,0.1l0.2,-0.1l-0.1,-0.2C126.8,365.2 123.7,365 120.6,365" />
-  <path
-      android:fillColor="#7FC2F8"
-      android:pathData="M240.1,356.4C240.1,356.4 240.1,356.4 240.1,356.4c-5.4,0.1 -10.7,1.1 -15.7,3l-0.1,0.1l0.2,0.2c3.5,-0.3 6.9,-0.4 10.4,-0.4c1.7,0 3.5,0 5.2,0.1c5.2,0.1 10.3,0.7 15.3,1.7c5.1,1.1 10,2.8 14.7,5.1c3.3,1.6 6.4,3.4 9.5,5.4c0.3,-0.1 0.6,-0.1 0.9,-0.2c-3,-2.5 -6.3,-4.7 -9.7,-6.6c-4.7,-2.6 -9.6,-4.6 -14.8,-6.1C250.9,357.2 245.5,356.4 240.1,356.4" />
-  <path
-      android:fillColor="#FF7FAB"
-      android:pathData="M358,519.9c-58.3,17.5 -119.4,15.2 -182.5,-2.7c-31.1,-8.8 -52.7,-37.1 -52.7,-69.5c0,-41.6 35.1,-74.8 76.7,-72c34.7,2.3 69.5,-0.7 103.2,-8.8c54.4,-13 106.6,28 106.6,83.8C409.4,482.6 388.5,510.7 358,519.9" />
-  <path
-      android:fillColor="#FB9DBE"
-      android:pathData="M323.7,377.1c-7.5,0 -15.1,1 -22.4,3l0,0l-0.1,0.2l0.2,0.1c5.9,-1.2 11.9,-1.8 17.9,-1.8c0.9,0 1.8,0 2.7,0c6.9,0.2 13.7,1.2 20.4,3c6.6,1.9 12.9,4.5 19,7.7c6.2,3.1 12.1,6.7 17.7,10.7h0.2v-0.2c-4.7,-5.3 -10.3,-9.7 -16.5,-13.1c-6.2,-3.4 -12.9,-5.9 -19.8,-7.3C336.6,377.8 330.2,377.1 323.7,377.1" />
-  <path
-      android:fillColor="#FB9DBE"
-      android:pathData="M191.8,387.8c-2.3,0 -4.6,0.2 -6.9,0.5c-5.6,0.9 -11.1,2.5 -16.2,5c-5.1,2.3 -10,5.2 -14.4,8.7c-4.4,3.5 -8.4,7.4 -11.9,11.8l0,0.2l0,0l0.2,0c3.8,-4.1 8,-7.7 12.5,-10.8c4.5,-3.1 9.4,-5.7 14.5,-7.7c5.1,-1.9 10.4,-3.3 15.7,-4.3c5.4,-1 10.9,-1.7 16.3,-2.1l0.1,-0.1l0,0l-0.1,-0.2C198.5,388.2 195.2,387.8 191.8,387.8" />
+    <path
+        android:fillColor="#C2CACC"
+        android:pathData="M451.5,482.8c0,-17 -41.2,-30.8 -92.1,-30.8c-9.3,0 -18.3,0.5 -26.8,1.3c-19.6,-3.3 -41.8,-5.2 -65.3,-5.2c-8,0 -15.9,0.2 -23.6,0.6c-21,-5.2 -46.9,-8.2 -75,-8.2c-70.2,0 -127,19 -127,42.5c0,21.2 46.3,38.7 106.8,42c26.4,12.4 69.8,20.5 118.8,20.5c68.1,0 125.2,-15.6 141.2,-36.8C434.3,503.4 451.5,493.8 451.5,482.8z" />
+    <path
+        android:fillColor="#FF9800"
+        android:pathData="M393.2,495.4c-46.7,-9.5 -87.9,-33.6 -124.8,-69.3c-18.2,-17.5 -22.6,-45 -10.7,-67.3c15.4,-28.7 51.9,-38.6 79.4,-21.3c23,14.4 48.1,25.2 74.4,32.1c42.3,11.2 63.1,58.7 42.5,97.2C442.4,488.7 417.6,500.4 393.2,495.4" />
+    <path
+        android:fillColor="#FFBC41"
+        android:pathData="M405.7,378.1l-0.2,0.1l0.1,0.2c5.2,1.6 10.2,3.8 14.8,6.5c4.7,2.7 9,5.9 12.8,9.7c3.7,3.8 7.2,7.9 10.2,12.3c3.1,4.4 6,9 8.5,13.8l0.2,0.1l0.1,-0.2c-1.2,-5.4 -3.3,-10.6 -6.3,-15.3c-3,-4.7 -6.7,-8.9 -11,-12.4c-4.2,-3.6 -8.8,-6.7 -13.8,-9.2C416.3,381.2 411.1,379.3 405.7,378.1L405.7,378.1" />
+    <path
+        android:fillColor="#FFBC41"
+        android:pathData="M306.2,338c-3.3,0 -6.6,0.3 -9.9,0.9c-4.3,0.8 -8.5,2.1 -12.5,3.9l0,0l-0.1,0.2l0.2,0.1c4.1,-1.4 8.3,-2.3 12.6,-2.7c1.9,-0.2 3.8,-0.3 5.7,-0.3c2.3,0 4.7,0.1 7,0.4c4.2,0.7 8.3,1.6 12.3,2.9c4.1,1.2 8,2.7 12.1,4.3h0.2l0,0l0,-0.2c-3.3,-3 -7.2,-5.3 -11.3,-6.9c-4.2,-1.5 -8.6,-2.3 -13.1,-2.5C308.4,338 307.3,338 306.2,338" />
+    <path
+        android:fillColor="#42A5F5"
+        android:pathData="M96.9,504.7c57.2,8 114,-3 170.6,-28.8c27.9,-12.7 44,-42.4 39.4,-72.7c-6,-39 -43.7,-65 -82.2,-56.4c-32.1,7.2 -65.2,9.4 -98,6.6c-52.8,-4.3 -95.9,41.6 -87.8,93.9C43.4,477.2 67,500.5 96.9,504.7" />
+    <path
+        android:fillColor="#7FC2F8"
+        android:pathData="M120.6,365c-10.2,0 -20.4,1.9 -29.9,5.8c-6.3,2.4 -12.2,5.6 -17.6,9.7c-5.3,4.1 -9.9,9.1 -13.5,14.7v0.2h0.2c4.7,-4.5 9.8,-8.7 15.1,-12.5c5.2,-3.8 10.8,-7.1 16.7,-9.9c6,-2.7 12.2,-4.6 18.6,-5.8c5.2,-1 10.4,-1.4 15.6,-1.4c1.3,0 2.7,0 4,0.1l0.2,-0.1l-0.1,-0.2C126.8,365.2 123.7,365 120.6,365" />
+    <path
+        android:fillColor="#7FC2F8"
+        android:pathData="M240.1,356.4C240.1,356.4 240.1,356.4 240.1,356.4c-5.4,0.1 -10.7,1.1 -15.7,3l-0.1,0.1l0.2,0.2c3.5,-0.3 6.9,-0.4 10.4,-0.4c1.7,0 3.5,0 5.2,0.1c5.2,0.1 10.3,0.7 15.3,1.7c5.1,1.1 10,2.8 14.7,5.1c3.3,1.6 6.4,3.4 9.5,5.4c0.3,-0.1 0.6,-0.1 0.9,-0.2c-3,-2.5 -6.3,-4.7 -9.7,-6.6c-4.7,-2.6 -9.6,-4.6 -14.8,-6.1C250.9,357.2 245.5,356.4 240.1,356.4" />
+    <path
+        android:fillColor="#FF7FAB"
+        android:pathData="M358,519.9c-58.3,17.5 -119.4,15.2 -182.5,-2.7c-31.1,-8.8 -52.7,-37.1 -52.7,-69.5c0,-41.6 35.1,-74.8 76.7,-72c34.7,2.3 69.5,-0.7 103.2,-8.8c54.4,-13 106.6,28 106.6,83.8C409.4,482.6 388.5,510.7 358,519.9" />
+    <path
+        android:fillColor="#FB9DBE"
+        android:pathData="M323.7,377.1c-7.5,0 -15.1,1 -22.4,3l0,0l-0.1,0.2l0.2,0.1c5.9,-1.2 11.9,-1.8 17.9,-1.8c0.9,0 1.8,0 2.7,0c6.9,0.2 13.7,1.2 20.4,3c6.6,1.9 12.9,4.5 19,7.7c6.2,3.1 12.1,6.7 17.7,10.7h0.2v-0.2c-4.7,-5.3 -10.3,-9.7 -16.5,-13.1c-6.2,-3.4 -12.9,-5.9 -19.8,-7.3C336.6,377.8 330.2,377.1 323.7,377.1" />
+    <path
+        android:fillColor="#FB9DBE"
+        android:pathData="M191.8,387.8c-2.3,0 -4.6,0.2 -6.9,0.5c-5.6,0.9 -11.1,2.5 -16.2,5c-5.1,2.3 -10,5.2 -14.4,8.7c-4.4,3.5 -8.4,7.4 -11.9,11.8l0,0.2l0,0l0.2,0c3.8,-4.1 8,-7.7 12.5,-10.8c4.5,-3.1 9.4,-5.7 14.5,-7.7c5.1,-1.9 10.4,-3.3 15.7,-4.3c5.4,-1 10.9,-1.7 16.3,-2.1l0.1,-0.1l0,0l-0.1,-0.2C198.5,388.2 195.2,387.8 191.8,387.8" />
 </vector>
