+++ Udacity/DessertPusher/app/src/main/res/drawable/marshmallow.xml
@@ -19,28 +19,28 @@
     android:height="700dp"
     android:viewportWidth="500"
     android:viewportHeight="700">
-  <path
-      android:fillColor="#C2CACC"
-      android:pathData="M167.8,451.8c-50.3,0 -91,13.7 -91,30.7c0,17 40.8,30.7 91,30.7c1.2,0 2.4,0 3.6,0c-3.8,3.4 -5.9,7.1 -5.9,10.9c0,17 40.8,30.7 91,30.7c50.3,0 91,-13.7 91,-30.7c0,-17 -40.8,-30.7 -91,-30.7c-1.2,0 -2.4,0 -3.6,0c3.8,-3.4 5.9,-7.1 5.9,-10.9C258.8,465.5 218,451.8 167.8,451.8" />
-  <path
-      android:fillColor="#C2CACC"
-      android:pathData="M354,430.5c-38.3,0 -69.3,13.7 -69.3,30.7c0,17 31,30.7 69.3,30.7s69.3,-13.7 69.3,-30.7C423.3,444.2 392.2,430.5 354,430.5" />
-  <path
-      android:fillColor="#90CAF9"
-      android:pathData="M413.4,462.3c0,17.4 -15.3,22.2 -65.7,22.2c-36.3,0 -65.7,-8 -65.7,-22.2c7.4,-64.3 -2.6,-107.9 0,-125.3h131.5C415.8,384 409.6,428.7 413.4,462.3z" />
-  <path
-      android:fillColor="#BBDEFB"
-      android:pathData="M413.4,337c0,8.5 -33.8,26.6 -70.1,26.6s-61.4,-18 -61.4,-26.6c0,-8.5 34.3,-11.6 70.6,-11.6S413.4,328.5 413.4,337z" />
-  <path
-      android:fillColor="#F48FB1"
-      android:pathData="M224.2,480.6c0,18.5 -16.2,23.6 -69.8,23.6c-38.6,0 -69.8,-8.5 -69.8,-23.6c7.8,-68.3 -2.8,-114.6 0,-133h139.6C226.7,397.5 220.2,444.9 224.2,480.6z" />
-  <path
-      android:fillColor="#F8BBD0"
-      android:pathData="M224.2,347.6c0,9.1 -35.8,28.2 -74.4,28.2s-65.2,-19.2 -65.2,-28.2c0,-9.1 36.4,-12.3 75,-12.3S224.2,338.5 224.2,347.6z" />
-  <path
-      android:fillColor="#E7E7E7"
-      android:pathData="M337.8,518.7c0,21.6 -19,27.6 -81.7,27.6c-45.1,0 -81.7,-10 -81.7,-27.6c9.2,-79.9 -3.3,-134.1 0,-155.7h163.4C340.7,421.4 333.1,476.9 337.8,518.7z" />
-  <path
-      android:fillColor="#F5F5F5"
-      android:pathData="M337.8,363c0,10.6 -41.9,33 -87.1,33s-76.3,-22.4 -76.3,-33s42.6,-14.4 87.8,-14.4S337.8,352.4 337.8,363z" />
+    <path
+        android:fillColor="#C2CACC"
+        android:pathData="M167.8,451.8c-50.3,0 -91,13.7 -91,30.7c0,17 40.8,30.7 91,30.7c1.2,0 2.4,0 3.6,0c-3.8,3.4 -5.9,7.1 -5.9,10.9c0,17 40.8,30.7 91,30.7c50.3,0 91,-13.7 91,-30.7c0,-17 -40.8,-30.7 -91,-30.7c-1.2,0 -2.4,0 -3.6,0c3.8,-3.4 5.9,-7.1 5.9,-10.9C258.8,465.5 218,451.8 167.8,451.8" />
+    <path
+        android:fillColor="#C2CACC"
+        android:pathData="M354,430.5c-38.3,0 -69.3,13.7 -69.3,30.7c0,17 31,30.7 69.3,30.7s69.3,-13.7 69.3,-30.7C423.3,444.2 392.2,430.5 354,430.5" />
+    <path
+        android:fillColor="#90CAF9"
+        android:pathData="M413.4,462.3c0,17.4 -15.3,22.2 -65.7,22.2c-36.3,0 -65.7,-8 -65.7,-22.2c7.4,-64.3 -2.6,-107.9 0,-125.3h131.5C415.8,384 409.6,428.7 413.4,462.3z" />
+    <path
+        android:fillColor="#BBDEFB"
+        android:pathData="M413.4,337c0,8.5 -33.8,26.6 -70.1,26.6s-61.4,-18 -61.4,-26.6c0,-8.5 34.3,-11.6 70.6,-11.6S413.4,328.5 413.4,337z" />
+    <path
+        android:fillColor="#F48FB1"
+        android:pathData="M224.2,480.6c0,18.5 -16.2,23.6 -69.8,23.6c-38.6,0 -69.8,-8.5 -69.8,-23.6c7.8,-68.3 -2.8,-114.6 0,-133h139.6C226.7,397.5 220.2,444.9 224.2,480.6z" />
+    <path
+        android:fillColor="#F8BBD0"
+        android:pathData="M224.2,347.6c0,9.1 -35.8,28.2 -74.4,28.2s-65.2,-19.2 -65.2,-28.2c0,-9.1 36.4,-12.3 75,-12.3S224.2,338.5 224.2,347.6z" />
+    <path
+        android:fillColor="#E7E7E7"
+        android:pathData="M337.8,518.7c0,21.6 -19,27.6 -81.7,27.6c-45.1,0 -81.7,-10 -81.7,-27.6c9.2,-79.9 -3.3,-134.1 0,-155.7h163.4C340.7,421.4 333.1,476.9 337.8,518.7z" />
+    <path
+        android:fillColor="#F5F5F5"
+        android:pathData="M337.8,363c0,10.6 -41.9,33 -87.1,33s-76.3,-22.4 -76.3,-33s42.6,-14.4 87.8,-14.4S337.8,352.4 337.8,363z" />
 </vector>
