+++ Udacity/DessertPusher/app/src/main/res/layout/activity_main.xml
@@ -32,7 +32,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context="com.nudennie.dessertpusher.MainActivity">
 
         <ImageView
             android:id="@+id/background_image"
