--- Udacity/DiceRoller/app/src/main/java/com/tc2r1/diceroller/MainActivity.kt
@@ -1,33 +1,0 @@
-package com.tc2r1.diceroller
-
-import android.os.Bundle
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_main.*
-import java.util.*
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        btn_roll.setOnClickListener {
-            Toast.makeText(this, "Rolling Dice", Toast.LENGTH_SHORT).show()
-            rollDice()
-        }
-    }
-
-    private fun rollDice() {
-        val drawableResource = when (Random().nextInt(6) + 1) {
-            1 -> R.drawable.dice_1
-            2 -> R.drawable.dice_2
-            3 -> R.drawable.dice_3
-            4 -> R.drawable.dice_4
-            5 -> R.drawable.dice_5
-            else -> R.drawable.dice_6
-        }
-        iv_dice.setImageResource(drawableResource)
-    }
-
-}
