--- Udacity/TriviaApp/app/src/main/java/com/example/android/navigation/GameWonFragment.kt
@@ -1,86 +1,0 @@
-/*
- * Copyright 2018, The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.android.navigation
-
-import android.content.Intent
-import android.os.Bundle
-import android.view.*
-import androidx.core.app.ShareCompat
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.navigation.findNavController
-import androidx.navigation.fragment.navArgs
-import com.example.android.navigation.databinding.FragmentGameWonBinding
-
-
-class GameWonFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        val binding: FragmentGameWonBinding = DataBindingUtil.inflate(
-            inflater, R.layout.fragment_game_won, container, false
-        )
-
-        binding.apply {
-            nextMatchButton.setOnClickListener { view: View ->
-                view.findNavController()
-                    .navigate(GameWonFragmentDirections.actionGameWonFragmentToTitleFragment())
-            }
-        }
-
-        setHasOptionsMenu(true)
-        return binding.root
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-        inflater.inflate(R.menu.winner_menu, menu)
-
-        // check if the activity resolves
-        getShareIntent().resolveActivity(requireActivity().packageManager).let {
-            // hide the share menu item.
-            menu.findItem(R.id.share)?.setVisible(false)
-        }
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.share -> shareSuccess()
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    private fun getShareIntent(): Intent {
-        val args: GameWonFragmentArgs by navArgs()
-
-        // Alternative way of creating intent for sharing.
-//        val shareIntent = Intent(Intent.ACTION_SEND)
-//        shareIntent.type = "text/plain"
-//        shareIntent.putExtra(Intent.EXTRA_TEXT, getString(R.string.share_success_text, args.numCorrect, args.numQuestions))
-//        return shareIntent
-        return ShareCompat.IntentBuilder.from(requireActivity())
-            .setText(getString(R.string.share_success_text, args.numCorrect, args.numQuestions))
-            .setText("text/plain")
-            .intent
-    }
-
-    private fun shareSuccess() {
-        startActivity(getShareIntent())
-
-    }
-}
