--- Udacity/TriviaApp/app/src/main/java/com/example/android/navigation/TitleFragment.kt
@@ -1,42 +1,0 @@
-package com.example.android.navigation
-
-import android.os.Bundle
-import android.view.*
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.navigation.findNavController
-import androidx.navigation.ui.NavigationUI
-import com.example.android.navigation.databinding.FragmentTitleBinding
-
-/**
- * A simple [Fragment] subclass.
- * Use the [TitleFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-class TitleFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        val binding: FragmentTitleBinding =
-            DataBindingUtil.inflate(inflater, R.layout.fragment_title, container, false)
-        binding.apply {
-            playButton.setOnClickListener { view: View ->
-                view.findNavController()
-                    .navigate(TitleFragmentDirections.actionTitleFragmentToGameFragment())
-            }
-        }
-        setHasOptionsMenu(true)
-        return binding.root
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        inflater.inflate(R.menu.overflow_menu, menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return NavigationUI.onNavDestinationSelected(item, view!!.findNavController())
-                || super.onOptionsItemSelected(item)
-    }
-}
