+++ Udacity/TriviaApp/app/src/main/res/layout/fragment_game_over.xml
@@ -17,7 +17,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.android.navigation.GameOverFragment">
+    tools:context="com.nudennie.navigation.com.nudennie.navigation.GameOverFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/gameOverConstraintLayout"
