+++ Udacity/TriviaApp/app/src/main/res/layout/fragment_rules.xml
@@ -20,7 +20,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context="com.example.android.navigation.RulesFragment">
+    tools:context="com.nudennie.navigation.com.nudennie.navigation.RulesFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
