+++ Udacity/TriviaApp/app/src/main/res/navigation/navigation.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/titleFragment"
-        android:name="com.example.android.navigation.TitleFragment"
+        android:name="com.nudennie.navigation.TitleFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_title">
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/gameFragment"
-        android:name="com.example.android.navigation.GameFragment"
+        android:name="com.nudennie.navigation.GameFragment"
         android:label="@string/title_android_trivia_question"
         tools:layout="@layout/fragment_game">
         <action
@@ -44,7 +44,7 @@
     </fragment>
     <fragment
         android:id="@+id/gameOverFragment"
-        android:name="com.example.android.navigation.GameOverFragment"
+        android:name="com.nudennie.navigation.GameOverFragment"
         android:label="@string/game_over"
         tools:layout="@layout/fragment_game_over">
         <action
@@ -58,7 +58,7 @@
     </fragment>
     <fragment
         android:id="@+id/gameWonFragment"
-        android:name="com.example.android.navigation.GameWonFragment"
+        android:name="com.nudennie.navigation.GameWonFragment"
         android:label="@string/congratulations"
         tools:layout="@layout/fragment_game_won">
         <action
@@ -78,12 +78,12 @@
     </fragment>
     <fragment
         android:id="@+id/aboutFragment"
-        android:name="com.example.android.navigation.AboutFragment"
+        android:name="com.nudennie.navigation.AboutFragment"
         android:label="@string/title_about_trivia"
         tools:layout="@layout/fragment_about" />
     <fragment
         android:id="@+id/rulesFragment"
-        android:name="com.example.android.navigation.RulesFragment"
+        android:name="com.nudennie.navigation.RulesFragment"
         android:label="fragment_rules"
         tools:layout="@layout/fragment_rules" />
 </navigation>
